Change tmdb lite ids to longs
This shouldn't affect the json parsing and makes it simpler working
with the database.

+++ app/src/test/java/com/rbraithwaite/untitledmovietracker/test_utils/TestDependencyManager.kt
@@ -150,10 +150,10 @@ class TestDependencyManager(
                 depManager.backend.insert(movie) { copy(id = it) }
             }
             for (tvShow in tvShows) {
-                depManager.backend.insert(tvShow) { copy(id = it.toInt()) }
+                depManager.backend.insert(tvShow) { copy(id = it) }
             }
             for (person in people) {
-                depManager.backend.insert(person) { copy(id = it.toInt()) }
+                depManager.backend.insert(person) { copy(id = it) }
             }
         }
     }
